requirements.txt
Pillow
PyOpenGL
PyOpenGL_accelerate
opensimplex